deepseek/requirements.txt
openai
python-dotenv
tiktoken
customtkinter